# src/main/java/com/ejemplo/demo/api/dto/CategoriaRequest.java
package com.ejemplo.demo.api.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record CategoriaRequest(
    @NotBlank(message = "El nombre de la categoria es obligatorio")
    @Size(max = 100, message = "El nombre no puede exceder los 100 caracteres")
    String nombre,
    
    @Size(max = 255)
    String descripcion
) {}