# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Workshop Spring Boot API
  version: 1.0.0
  description: Contrato publico del workshop Spring Boot 3.x.
  contact:
    name: Jorge Mario Cano Cobon
    email: jcanoc5@miumg.edu.gt
tags:
  - name: workshop
    description: Endpoints principales del taller
  - name: simulaciones
    description: Endpoints de simulaciones de negocio
  - name: demo-estado
    description: Comparacion de estado singleton y manual
  - name: categorias
    description: CRUD de categorias
  - name: productos
    description: CRUD de productos
paths:
  /api/v1:
    get:
      tags:
        - workshop
      operationId: getWorkshopHealth
      responses:
        '200':
          description: Estado del workshop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkshopHealthResponse'
  /api/v1/saludos:
    get:
      tags:
        - workshop
      operationId: saludarPorGet
      parameters:
        - name: nombre
          in: query
          required: false
          schema:
            type: string
            default: Mundo
      responses:
        '200':
          description: Saludo generado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaludoResponse'
    post:
      tags:
        - workshop
      operationId: saludarPorPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaludoRequest'
      responses:
        '200':
          description: Saludo generado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaludoResponse'
        '400':
          description: Error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/simulaciones/prestamo:
    post:
      tags:
        - simulaciones
      operationId: simularPrestamo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrestamoRequest'
      responses:
        '200':
          description: Resultado de la simulacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrestamoResponse'
        '400':
          description: Error de validacion o regla de negocio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/demo/estado/singleton/{valor}:
    post:
      tags:
        - demo-estado
      operationId: actualizarSingleton
      parameters:
        - $ref: '#/components/parameters/ValorPath'
      responses:
        '200':
          description: Estado singleton actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'
  /api/v1/demo/estado/singleton:
    get:
      tags:
        - demo-estado
      operationId: obtenerSingleton
      responses:
        '200':
          description: Estado singleton actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'
  /api/v1/demo/estado/singleton/reset:
    post:
      tags:
        - demo-estado
      operationId: reiniciarSingleton
      responses:
        '200':
          description: Estado singleton reiniciado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'
  /api/v1/demo/estado/manual/{valor}:
    post:
      tags:
        - demo-estado
      operationId: actualizarManual
      parameters:
        - $ref: '#/components/parameters/ValorPath'
      responses:
        '200':
          description: Estado manual actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'
  /api/v1/demo/estado/manual:
    get:
      tags:
        - demo-estado
      operationId: obtenerManual
      responses:
        '200':
          description: Estado manual actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'
  /api/v1/categorias:
    get:
      tags:
        - categorias
      operationId: listarCategorias
      responses:
        '200':
          description: Listado de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoriaResponse'
    post:
      tags:
        - categorias
      operationId: crearCategoria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaRequest'
      responses:
        '201':
          description: Categoria creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaResponse'
        '400':
          description: Error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/categorias/{id}:
    get:
      tags:
        - categorias
      operationId: obtenerCategoria
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Categoria encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaResponse'
        '404':
          description: Categoria no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - categorias
      operationId: actualizarCategoria
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaRequest'
      responses:
        '200':
          description: Categoria actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaResponse'
        '400':
          description: Error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Categoria no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - categorias
      operationId: eliminarCategoria
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Categoria eliminada
        '404':
          description: Categoria no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/productos:
    get:
      tags:
        - productos
      operationId: listarProductos
      responses:
        '200':
          description: Listado de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoResponse'
    post:
      tags:
        - productos
      operationId: crearProducto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoRequest'
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '400':
          description: Error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/productos/{id}:
    get:
      tags:
        - productos
      operationId: obtenerProducto
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - productos
      operationId: actualizarProducto
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoRequest'
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '400':
          description: Error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - productos
      operationId: eliminarProducto
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Producto eliminado
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ValorPath:
      name: valor
      in: path
      required: true
      schema:
        type: integer
        format: int32
  schemas:
    WorkshopHealthResponse:
      type: object
      required:
        - estado
        - mensaje
      properties:
        estado:
          type: string
          example: ok
        mensaje:
          type: string
          example: Workshop Spring Boot activo
    SaludoRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          minLength: 1
    SaludoResponse:
      type: object
      required:
        - mensaje
        - fecha
      properties:
        mensaje:
          type: string
        fecha:
          type: string
          format: date-time
    PrestamoRequest:
      type: object
      required:
        - monto
        - tasaAnual
        - meses
      properties:
        monto:
          type: number
          minimum: 0.01
        tasaAnual:
          type: number
          minimum: 0.01
        meses:
          type: integer
          format: int32
          minimum: 1
          maximum: 360
    PrestamoResponse:
      type: object
      required:
        - cuotaMensual
        - interesTotal
        - totalPagar
      properties:
        cuotaMensual:
          type: number
        interesTotal:
          type: number
        totalPagar:
          type: number
    EstadoResponse:
      type: object
      required:
        - tipo
        - valorActual
      properties:
        tipo:
          type: string
        valorActual:
          type: integer
          format: int32
    CategoriaRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        descripcion:
          type: string
          maxLength: 255
    CategoriaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        descripcion:
          type: string
    ProductoRequest:
      type: object
      required:
        - nombre
        - precio
        - categoriaId
      properties:
        nombre:
          type: string
          minLength: 1
        precio:
          type: number
          minimum: 0.01
        categoriaId:
          type: integer
          format: int64
    ProductoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        precio:
          type: number
        nombreCategoria:
          type: string
    ErrorResponse:
      type: object
      properties:
        codigo:
          type: string
        mensaje:
          type: string
        timestamp:
          type: string
          format: date-time
        detalles:
          type: object
          additionalProperties:
            type: string